# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export interface Service {
  id: number;
  slug: string;
  name: string;
  description: string;
  durationMinutes: number;
  priceCents: number;
  sortOrder: number;
  active: boolean;
}

export interface TimeSlot {
  /** HH:MM in 24-hour format */
  time: string;
  available: boolean;
}

export interface DayAvailability {
  /** YYYY-MM-DD */
  date: string;
  closed: boolean;
  slots: TimeSlot[];
}

export interface CreateBookingRequest {
  serviceId: number;
  /** YYYY-MM-DD */
  date: string;
  /** HH:MM */
  time: string;
  /** @minLength 1 */
  customerName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  vehicle: string;
  notes?: string;
}

export interface BlockedDate {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  reason: string;
  createdAt: string;
}

export interface CreateBlockedDateRequest {
  /** YYYY-MM-DD */
  date: string;
  reason?: string;
}

/**
 * At least one field must be provided.
 */
export interface UpdateBookingRequest {
  /** @minLength 1 */
  customerName?: string;
  email?: string;
  /** @minLength 7 */
  phone?: string;
  /** @minLength 1 */
  vehicle?: string;
  notes?: string;
}

export interface RescheduleBookingRequest {
  /** YYYY-MM-DD */
  date: string;
  /** HH:MM */
  time: string;
}

export interface RuleSlot {
  id: number;
  /** HH:MM */
  time: string;
}

export interface ServiceDayRule {
  id: number;
  serviceId: number;
  serviceSlug: string;
  serviceName: string;
  /**
   * 0=Sun..6=Sat
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  wholeDayLock: boolean;
  active: boolean;
  slots: RuleSlot[];
}

export interface CreateServiceRuleRequest {
  serviceId: number;
  /**
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  wholeDayLock: boolean;
  slots?: string[];
}

export interface UpdateServiceRuleRequest {
  wholeDayLock?: boolean;
  active?: boolean;
}

export interface AddRuleSlotRequest {
  /** HH:MM */
  time: string;
}

export interface Booking {
  id: number;
  serviceId: number;
  serviceName: string;
  servicePriceCents: number;
  serviceDurationMinutes: number;
  customerName: string;
  email: string;
  phone: string;
  vehicle: string;
  notes: string;
  scheduledAt: string;
  status: string;
  /** Returned to the booking creator and via the manage endpoints. Used in the customer self-manage URL. */
  manageToken?: string | null;
  createdAt: string;
}

export type GetAvailabilityParams = {
  /**
   * YYYY-MM-DD
   */
  startDate: string;
  /**
   * YYYY-MM-DD
   */
  endDate: string;
  /**
   * When provided, slots are filtered to those bookable for this service (e.g. Fridays only return slots for the Express service).
   */
  serviceId?: number;
};

export type GetManagedBookingParams = {
  token: string;
};

export type CancelManagedBookingParams = {
  token: string;
};

export type RescheduleManagedBookingParams = {
  token: string;
};