# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /booking/availability:
    get:
      operationId: getAvailability
      tags: [booking]
      summary: Get availability for a date range
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            description: YYYY-MM-DD
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            description: YYYY-MM-DD
        - name: serviceId
          in: query
          required: false
          schema:
            type: integer
            description: When provided, slots are filtered to those bookable for this service (e.g. Fridays only return slots for the Express service).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayAvailability"
  /booking/bookings:
    post:
      operationId: createBooking
      tags: [booking]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /booking/manage/{id}:
    get:
      operationId: getManagedBooking
      tags: [booking]
      summary: Get a booking via the customer's self-manage link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /booking/manage/{id}/cancel:
    post:
      operationId: cancelManagedBooking
      tags: [booking]
      summary: Cancel a booking via the customer's self-manage link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cancelled
        "400":
          description: Cannot cancel (e.g. already passed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /booking/manage/{id}/reschedule:
    post:
      operationId: rescheduleManagedBooking
      tags: [booking]
      summary: Reschedule a booking via the customer's self-manage link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBookingRequest"
      responses:
        "200":
          description: Rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Slot already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    TimeSlot:
      type: object
      properties:
        time:
          type: string
          description: "HH:MM in 24-hour format"
        available:
          type: boolean
      required:
        - time
        - available
    DayAvailability:
      type: object
      properties:
        date:
          type: string
          description: YYYY-MM-DD
        closed:
          type: boolean
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - date
        - closed
        - slots
    CreateBookingRequest:
      type: object
      properties:
        serviceId:
          type: integer
        date:
          type: string
          description: YYYY-MM-DD
        time:
          type: string
          description: "HH:MM"
        customerName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
        vehicle:
          type: string
          minLength: 1
        notes:
          type: string
      required:
        - serviceId
        - date
        - time
        - customerName
        - email
        - phone
        - vehicle
...
    RescheduleBookingRequest:
      type: object
      properties:
        date:
          type: string
          description: YYYY-MM-DD
        time:
          type: string
          description: "HH:MM"
      required:
        - date
        - time
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        servicePriceCents:
          type: integer
        serviceDurationMinutes:
          type: integer
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        vehicle:
          type: string
        notes:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
        manageToken:
          type: string
          nullable: true
          description: Returned to the booking creator and via the manage endpoints. Used in the customer self-manage URL.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - serviceName
        - servicePriceCents
        - serviceDurationMinutes
        - customerName
        - email
        - phone
        - vehicle
        - notes
        - scheduledAt
        - status
        - createdAt